refactor: file structure, remove any
