+++ src/components/Button/Button.style.ts
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create<any>({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    padding: 16,
+    borderRadius: 3,
+    borderColor: 'lightgrey',
+    justifyContent: 'center',
+  },
+});
