+++ src/components/CustomText/CustomText.style.ts
@@ -1,0 +1,5 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  text: {includeFontPadding: false},
+});
