+++ src/components/FirstPage/FirstPage.style.ts
@@ -1,0 +1,8 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {flexDirection: 'row', alignItems: 'center', padding: 12},
+  text: {fontSize: 14, marginLeft: 4},
+  leftButton: {padding: 4, flex: 1, marginRight: 6},
+  rightButton: {padding: 4, flex: 1, marginLeft: 6},
+});
