+++ src/components/Header/Header.style.ts
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    padding: 16,
+    backgroundColor: 'white',
+  },
+  flex1: {flex: 1},
+  rightFlex: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    flex: 1,
+  },
+  marginR24: {marginRight: 24},
+});
