+++ src/components/Icons/iconName.ts
@@ -1,0 +1,14 @@
+export const iconName = {
+  left: 'left',
+  link: 'link',
+  instagram: 'instagram',
+  bell: 'bell',
+  'three-dot': 'three-dot',
+  'right-fill': 'right-fill',
+  'right-outline': 'right-outline',
+  grid: 'grid',
+  question: 'question',
+  media: 'media',
+  bottom: 'bottom',
+  heart: 'heart',
+};
