+++ src/components/ImageRow/ImageRow.style.ts
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    backgroundColor: 'white',
+    flexWrap: 'wrap',
+    flexDirection: 'row',
+  },
+});
