+++ src/components/MenuButton/MenuButton.style.ts
@@ -1,0 +1,5 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  button: {flex: 1, paddingVertical: 8, alignItems: 'center'},
+});
