+++ src/components/ProfileSection/ProfileSection.style.ts
@@ -1,0 +1,31 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {alignItems: 'center', marginTop: 24},
+  image: {
+    marginBottom: 20,
+  },
+  row: {flexDirection: 'row', alignItems: 'center'},
+  divider: {height: 20, width: 0.7, backgroundColor: '#e3e3e4'},
+  buttonRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    width: '78%',
+    marginBottom: 20,
+  },
+  flex1: {flex: 1},
+  instagramButton: {padding: 6, marginLeft: 6},
+  bottomButton: {padding: 16, marginLeft: 6},
+  textCenter: {textAlign: 'center'},
+  link: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginVertical: 12,
+  },
+  marginL4: {marginLeft: 4},
+  qaContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 20,
+  },
+});
