+++ src/components/ProfileSection/ProfileSection.tsx
@@ -1,19 +1,23 @@
 import React from 'react';
-import Button from 'components/Button';
-import CustomText from 'components/CustomText';
-import Icons from 'components/Icons';
-import Stat from 'components/Stat';
-import {Image, StyleSheet, View} from 'react-native';
+import {CustomText} from 'components/CustomText/CustomText';
+import {Icons} from 'components/Icons/Icons';
+import {Stat} from 'components/Stat/Stat';
+import {Image, View} from 'react-native';
+import {Button} from 'components/Button/Button';
+import {styles} from './ProfileSection.style';
 
 const imageSize = 120;
-const ProfileSection = () => {
+export const ProfileSection = () => {
   return (
     <View style={styles.container}>
       <Image
         source={{
           uri: 'https://i.ibb.co/1zrXY93/2e8477cde07b9020c34f31d5250e8f3a-c5-720x720.jpg',
         }}
-        style={styles.image}
+        style={[
+          styles.image,
+          {height: imageSize, width: imageSize, borderRadius: imageSize / 2},
+        ]}
       />
       <CustomText weight="SemiBold">@wealth</CustomText>
       <View style={styles.row}>
@@ -63,38 +67,3 @@ const ProfileSection = () => {
     </View>
   );
 };
-
-export default ProfileSection;
-
-const styles = StyleSheet.create({
-  container: {alignItems: 'center', marginTop: 24},
-  image: {
-    height: imageSize,
-    width: imageSize,
-    borderRadius: imageSize / 2,
-    marginBottom: 20,
-  },
-  row: {flexDirection: 'row', alignItems: 'center'},
-  divider: {height: 20, width: 0.7, backgroundColor: '#e3e3e4'},
-  buttonRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    width: '78%',
-    marginBottom: 20,
-  },
-  flex1: {flex: 1},
-  instagramButton: {padding: 6, marginLeft: 6},
-  bottomButton: {padding: 16, marginLeft: 6},
-  textCenter: {textAlign: 'center'},
-  link: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginVertical: 12,
-  },
-  marginL4: {marginLeft: 4},
-  qaContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-});
