+++ src/components/SecondPage/SecondPage.style.ts
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    backgroundColor: 'white',
+    minHeight: 200,
+  },
+  boldText: {textAlign: 'center', marginTop: 66},
+  center: {textAlign: 'center'},
+});
