+++ src/components/Stat/Stat.style.ts
@@ -1,0 +1,6 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {alignItems: 'center', padding: 20},
+  marginB6: {marginBottom: 6},
+});
