+++ src/components/Stat/Stat.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
-import CustomText from 'components/CustomText';
-import {StyleSheet, View} from 'react-native';
+import {CustomText} from 'components/CustomText/CustomText';
+import {View} from 'react-native';
+import {styles} from './Stat.style';
 
 type StatProps = {
   name: string;
   statNumber: string;
 };
 
-const Stat: React.FC<StatProps> = ({name, statNumber}) => {
+export const Stat: React.FC<StatProps> = ({name, statNumber}) => {
   return (
     <View style={styles.container}>
       <CustomText weight="Bold" fontSize={19} style={styles.marginB6}>
@@ -19,10 +20,3 @@ const Stat: React.FC<StatProps> = ({name, statNumber}) => {
     </View>
   );
 };
-
-export default Stat;
-
-const styles = StyleSheet.create({
-  container: {alignItems: 'center', padding: 20},
-  marginB6: {marginBottom: 6},
-});
